requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
requests
PyYAML
joblib
